# apps/web/public/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BIKIE API",
    "version": "1.0.0",
    "description": "Stable v1 contract under `/api/*`. Flutter and web share this surface. `/api/v2` is reserved for approved breaking changes only (ADR-028). Schemas are incrementally detailed; treat `.docs/API.md` as the narrative source of truth."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin (Next.js Route Handlers)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Better Auth HTTP-only session cookie (web)."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Better Auth bearer token (Flutter / mobile). Header: Authorization: Bearer <token>."
      },
      "CronBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Cron routes: Authorization: Bearer <CRON_SECRET>."
      }
    },
...
  },
  "paths": {
...
    "/api/admin/partners": {
      "get": {
        "operationId": "get_api_admin_partners",
        "tags": [
          "admin"
        ],
        "summary": "GET /api/admin/partners",
        "description": "Source: `apps/web/app/api/admin/partners/route.ts`. Auth class: `admin`. Current stable surface is v1 under `/api/*` (no `/api/v2` without ADR approval). Narrative docs: `.docs/API.md`.",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "x-bikie-auth": "admin",
        "x-bikie-source": "apps/web/app/api/admin/partners/route.ts",
        "responses": {
          "200": {
            "description": "Success (shape documented in .docs/API.md for this route).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Validation / bad request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden / membership / role"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
    "/api/partner/dashboard": {
      "get": {
        "operationId": "get_api_partner_dashboard",
        "tags": [
          "partner"
        ],
        "summary": "GET /api/partner/dashboard",
        "description": "Source: `apps/web/app/api/partner/dashboard/route.ts`. Auth class: `partner`. Current stable surface is v1 under `/api/*` (no `/api/v2` without ADR approval). Narrative docs: `.docs/API.md`.",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "x-bikie-auth": "partner",
        "x-bikie-source": "apps/web/app/api/partner/dashboard/route.ts",
        "responses": {
          "200": {
            "description": "Success (shape documented in .docs/API.md for this route).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Validation / bad request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden / membership / role"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
    "/api/user/complete-phone-signup": {
      "patch": {
        "operationId": "patch_api_user_complete-phone-signup",
        "tags": [
          "other"
        ],
        "summary": "PATCH /api/user/complete-phone-signup",
        "description": "Source: `apps/web/app/api/user/complete-phone-signup/route.ts`. Auth class: `session`. Current stable surface is v1 under `/api/*` (no `/api/v2` without ADR approval). Narrative docs: `.docs/API.md`.",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "x-bikie-auth": "session",
        "x-bikie-source": "apps/web/app/api/user/complete-phone-signup/route.ts",
        "responses": {
          "200": {
            "description": "Success (shape documented in .docs/API.md for this route).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Validation / bad request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden / membership / role"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
  },
...
}